Rename county component submodule.
